--- shared/src/androidMain/kotlin/com/rootstrap/flowforms/shared/Platform.kt
@@ -1,7 +1,0 @@
-package com.rootstrap.flowforms.shared
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
