--- shared/src/commonMain/kotlin/com/rootstrap/flowforms/shared/CommonForm.kt
@@ -1,4 +1,0 @@
-package com.rootstrap.flowforms.shared
-
-class CommonForm {
-}
