--- shared/src/commonMain/kotlin/com/rootstrap/flowforms/shared/Platform.kt
@@ -1,7 +1,0 @@
-package com.rootstrap.flowforms.shared
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
