--- shared/src/iosMain/kotlin/com/rootstrap/flowforms/shared/Platform.kt
@@ -1,9 +1,0 @@
-package com.rootstrap.flowforms.shared
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
